Avoid legacy C++ toolchain and configuration APIs (#15)
* Update expected_file.json with bazel-bin iquotes

The compile_commands.json now includes
-iquote bazel-out/k8-fastbuild/bin
and
-iquote bazel-out/k8-fastbuild/bin/external/bazel_tools

This is likely due to difference between bazel versions.

* Avoid legacy C++ toolchain and configuration APIs

According to the Bazel 0.19.0 release notes, Bazel 0.20 introduces
incompatible changes:
https://docs.bazel.build/versions/master/skylark/backward-compatibility.html#disable-legacy-c-configuration-api
and
https://docs.bazel.build/versions/master/skylark/backward-compatibility.html#disable-legacy-c-toolchain-api

This change uses the examples from the docs to use different APIs and
builtins so that the compilation_database_aspect works with
--incompatible_disable_legacy_cpp_toolchain_skylark_api
and
--incompatible_disable_legacy_flags_cc_toolchain_api

This commit also changes the order of the flags in expected_file.json to
match the new order of output.

Test: Ran generate.sh on my bazel project before and after the change,
and all the compile_commands.json files contained the same flags (though
in a different order). Note that I'm not on a mac so one code path was
not explicitly tested.