Corrections to test runner
We were failing early on migration issues, but we want to run all the
tests first without any migration checks.

Once we actually started checking the diffs, we noticed that they were
out of date. Now updated.

The ycm output check was only runnable on Travis so that has been
changed to be runnable on dev machines.