allow generate.sh to be present at any filepath
Similar to the logic we follow in the YCM plugin, we can configure
generate.sh to ask bazel to use a one-time repository location for
getting the aspects file. This removes the requirement for
generate.sh to be present in the user's workspace.

This was not an option before because the override_repository flag
was not available.

This is also a good opportunity to spruce the README with revised and
simplified instructions.