Remove windows no-op (#66)
Small change to remove the hard stop for windows. For those on windows running bazel with `--compiler=clang-cl` the aspects seem to run fine.

Merging this in for others might be a problem if they are expecting the no-op. It would require others to using the `disable` attribute. Which I don't think is too much to ask since otherwise this is a hard stop for clang-cl users.