Enable passing options on to Bazel from generate.sh (#82)
Passing options was allowed before but the introduction of getopts
interfered by trying to parse the bazel options.

The most conventional way here is to separate the script options and
bazel options with `--`, but we were currently not splitting the args
vector like that. This change starts doing this.

README.md has been updated.