generate.sh: match target kind from the start of the string (#83)
In my project I use [emsdk](https://github.com/emscripten-core/emsdk/tree/main/bazel) and have `wasm_cc_binary` targets. These targets don't build with the standard config, thus when bazel-compdb matches against them it fails. The easy fix is to just change the pattern so that it requires `cc_` or `objc_` to be at the start of the string.

Co-authored-by: Tom Birch <tom@powerlinespro.com>