Also propagate through filegroup type rules
This will allow people to define the aspect only on a single filegroup
like target, instead of replicating the cc targets list in both the
filegroup and the compilation_database rules.

We support propagating through the attributes srcs and deps, which
should cover filegroup and pkg_tar at the very least.

Fixes #84.